ref: Move nameTrial into its own file
