+++ src/trials/name.js
@@ -1,0 +1,14 @@
+import htmlKeyboardResponse from "@jspsych/plugin-html-keyboard-response";
+
+import { LANGUAGE } from "../config/main";
+import { h1 } from "../lib/markup/tags";
+
+/** Task that displays the name of the experiment */
+const nameTrial = {
+  type: htmlKeyboardResponse,
+  stimulus: h1(LANGUAGE.name),
+  choices: "NO_KEYS",
+  trial_duration: 1000,
+};
+
+export { nameTrial };
